+++ gen_docs.sh
@@ -1,3 +1,4 @@
-#/usr/bin/bash
-
-terraform-docs markdown ./
+#!/usr/bin/env bash
+terraform-docs markdown table --hide-empty --output-file README.md --sort .
+terraform-docs markdown table --hide-empty --output-file README.md --sort ./examples/basic
+terraform-docs markdown table --hide-empty --output-file README.md --sort ./examples/rsc_managed_exocompute